refactor: Change from int to uint8 to make it more performant/less memory intensive

+++ src/game_of_life/model.py
@@ -3,7 +3,7 @@ from typing import ClassVar
 import numpy as np
 import numpy.typing as npt
 
-NDArrayInt = npt.NDArray[np.int_]
+NDArrayInt = npt.NDArray[np.uint8]
 
 
 class Grid:
